requirements.txt
numpy
pandas
scipy
wheel
chaospy==4.2.4
SALib
pytest
pytest-pep8
pytest-benchmark
pytest-dependency
SQLAlchemy
jsonpickle
cerberus
dask[complete]
dask_jobqueue
cloudpickle
scikit-learn
jinja2
kubernetes
autopep8
squarify
dill
qcg-pilotjob
